requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
absl-py=0.8.1=pypi_0
asn1crypto=0.24.0=py36_0
atomicwrites=1.3.0=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=py36_0
beautifulsoup4=4.7.1=py36_1
blas=1.0=mkl
bzip2=1.0.6=h14c3975_5
ca-certificates=2019.1.23=0
cachetools=3.1.1=pypi_0
certifi=2019.3.9=py36_0
cffi=1.12.3=py36h2e261b9_0
chardet=3.0.4=py36_1
cloudpickle=1.2.2=pypi_0
conda=4.6.14=py36_0
conda-build=3.17.8=py36_0
cryptography=2.6.1=py36h1ba5d50_0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
cython=0.29.14=pypi_0
decorator=4.4.0=py36_1
filelock=3.0.10=py36_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
glob2=0.6=py36_1
google-auth=1.7.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.25.0=pypi_0
gym=0.15.4=pypi_0
h5py=2.10.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.8=py36_0
imageio=2.6.1=pypi_0
importlib-metadata=0.23=pypi_0
intel-openmp=2019.3=199
ipython=7.5.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
isodate=0.6.0=pypi_0
jedi=0.13.3=py36_0
jinja2=2.10.1=py36_0
joblib=0.14.0=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=pypi_0
libarchive=3.3.3=h5d8350f_5
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
liblief=0.9.0=h7725739_2
libpng=1.6.37=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libxml2=2.9.9=he19cac6_0
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h49e0be7_2
markdown=3.1.1=pypi_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.1=pypi_0
mkl=2019.3=199
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
more-itertools=7.2.0=pypi_0
natsort=6.2.0=pypi_0
ncurses=6.1=he6710b0_1
networkx=2.4=pypi_0
ninja=1.9.0=py36hfd86e86_0
nose=1.3.7=pypi_0
numpy=1.16.3=py36h7e9f1db_0
numpy-base=1.16.3=py36hde5b4d6_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
opencv-python=4.1.1.26=pypi_0
openssl=1.1.1b=h7b6447c_1
packaging=19.2=pypi_0
pandas=0.25.3=pypi_0
parso=0.4.0=py_0
patchelf=0.9=he6710b0_3
pexpect=4.7.0=py36_0
phyre=0.2.1=pypi_0
pickleshare=0.7.5=py36_0
pillow=6.0.0=py36h34e0f95_0
pip=19.1=py36_0
pkginfo=1.5.0.1=py36_0
pluggy=0.13.0=pypi_0
plyfile=0.7.1=pypi_0
progressbar2=3.47.0=pypi_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.10.0=pypi_0
psutil=5.6.2=py36h7b6447c_0
ptyprocess=0.6.0=py36_0
py=1.8.0=pypi_0
py-lief=0.9.0=py36h7725739_2
pyasn1=0.4.7=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycosat=0.6.3=py36h14c3975_0
pycparser=2.19=py36_0
pyglet=1.3.2=pypi_0
pygments=2.3.1=py36_0
pymunk=5.6.0=pypi_0
pyopenssl=19.0.0=py36_0
pyparsing=2.4.5=pypi_0
pysocks=1.6.8=py36_0
pytest=5.2.4=pypi_0
python=3.6.8=h0371630_0
python-dateutil=2.8.1=pypi_0
python-libarchive-c=2.8=py36_6
python-utils=2.3.0=pypi_0
pytorch=1.1.0=py3.6_cuda10.0.130_cudnn7.5.1_0
pytz=2019.1=py_0
pyyaml=5.1=py36h7b6447c_0
rdflib=4.2.2=pypi_0
readline=7.0=h7b6447c_5
requests=2.21.0=py36_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
ruamel_yaml=0.15.46=py36h14c3975_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.2=pypi_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_0
soupsieve=1.8=py36_0
sqlite=3.27.2=h7b6447c_0
tabulate=0.8.5=pypi_0
tensorboard=2.0.1=pypi_0
tensorboard-logger=0.1.0=pypi_0
thrift=0.11.0=pypi_0
tk=8.6.8=hbc83047_0
torch-cluster=1.4.5=pypi_0
torch-geometric=1.3.2=pypi_0
torch-scatter=1.4.0=pypi_0
torch-sparse=0.4.3=pypi_0
torch-spline-conv=1.1.1=pypi_0
torchvision=0.2.2=py_3
tornado=6.0.3=pypi_0
tqdm=4.31.1=py36_1
traitlets=4.3.2=py36_0
urllib3=1.24.2=py36_0
wcwidth=0.1.7=py36_0
werkzeug=0.16.0=pypi_0
wheel=0.33.1=py36_0
xz=5.2.4=h14c3975_4
yacs=0.1.6=pypi_0
yaml=0.1.7=had09818_2
zipp=0.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
